requirements.txt
numpy
pandas
matplotlib
seaborn
opencv-python-headless
streamlit
scipy
tensorflow
pillow
ipykernel
